インターバルタイマーの処理をシーン:Playに移動!!
